backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
djoser==2.2.2
djangorestframework-simplejwt==5.3.1
django-filter==23.5
pytest-django==4.8.0
gunicorn==21.2.0
whitenoise==6.6.0
celery==5.3.6
redis==5.0.1
PyJWT==2.8.0
requests==2.31.0
Pillow==10.2.0
django-storages==1.14.2
python-dotenv==1.0.0
nltk==3.8.1
django-oauth-toolkit==2.3.0
coreapi==2.3.3
pyyaml==6.0.1
uritemplate==4.1.1
drf-yasg==1.21.7
social-auth-app-django==5.4.0
dj-database-url==2.1.0
boto3==1.34.34
django-celery-beat==2.5.0
django-celery-results==2.5.1
sentry-sdk==1.39.1 